0.9.0 release. Loads of changes:
-v for version
-u cleaner usage output
-i gives basic informations about todo database
-h -> usage

~/.todosqliterc appears with default SQLite output tweaks.
~/.todorc is smaller.
Support is limited to standard SQL and FTS3 if available.
Exits properly if db uses FTS and SQLite does not support it.
Exits properly if too much args are given instead of silently ignoring them.
Database tables format is normaly set in stone. todo version is stored so later changes and conversion are possible.